Real businesses don't have random customers. They have patterns.
The first point to work is it is a practice exercise and the lack of data is a risk.
When I say risk I'm talking about how to delivery value if we have no real situation to work.

So, I decide to go with a demand generator and this commit is the first one of many exercises
I plan to run toward a good solution for this problema.

You can check the employee archetype idea babe steps too. There is just one parameter(skill_level)
 but I plan plenty more parameters to the employees archetype.

The commit title is from Gemini. It is there because sound great.