Real businesses don't have random customers. They have patterns.
Start by understanding the use case.

- The first point to work is it is a practice exercise and the lack of data is a risk.
When I say risk I'm talking about how to delivery value if we have no real situation to work.

- Realistic Data
Work with realistic data is one of the main rules for this project and beacuse
this, instead of go direct to the algo we decide to work at 'scripting ideas' and
the first one is the greedy scheduler just to set a baseline for the expeirment.

So, I decide to go with a demand generator and this commit is the first one of many exercises
I plan to run toward a good solution for this problema.

- Employee representation need attributes.
You can check the employee archetype idea babe steps too. There is just one parameter(skill_level)
 but I plan plenty more parameters to the employees archetype.

- Have a Makefile and ignore bin / db file and vim swap files.

- document decisions into Engineer Docs.