Product CheckPoint
This is the moment where we stop and look for the status of the product.

We realize that follow the original plan and go strict with it by focus on
unestructured data entry delivery the wrong value because the is a prior barrier
to use the product that is **can see the schedule** which means: delivery the first
value of the product before go with features.