Export the roster as csv
- Export roster as csv
With this capability we can run an analise of the result and define the next step.

The csv looks like:

Hour,Employee Name,Role,Hourly Rate,Is Safety Senior?
08:00,Alice (Vet),Senior,50.00,YES
08:00,Dave (Jun),Junior,20.00,No
08:00,Frank (Jun),Junior,21.00,No
09:00,Alice (Vet),Senior,50.00,YES

and have evidence of a problem in our algorithm.

Cost Optimization != Operational Feasibility

Sometimes an employee is working at 9am, then not working at 10am, and come back
to work at 11am. A shift need to be at least 4 hours straight.

- have a shift summary.

- update Makefile to run shiftopt and export csv